17 спринт
почистил всякого, проверил всё по критериям успешного выполнения задания
