+++ app/src/main/java/com/example/playlistmaker/data/search/HistoryStorageRepo.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain.search
+package com.example.playlistmaker.data.search
 
 import com.example.playlistmaker.domain.models.Track
 import com.example.playlistmaker.domain.utils.Resource
