+++ app/src/main/java/com/example/playlistmaker/data/settings/SettingsRepo.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain.settings
+package com.example.playlistmaker.data.settings
 
 
 import com.example.playlistmaker.domain.settings.model.ThemeList
