+++ app/src/main/java/com/example/playlistmaker/data/settings/impl/SettingsRepoImpl.kt
@@ -1,42 +1,43 @@
 package com.example.playlistmaker.data.settings.impl
 
 import android.content.SharedPreferences
-import android.util.Log
-import com.example.playlistmaker.domain.settings.SettingsRepo
+import androidx.appcompat.app.AppCompatDelegate
+import com.example.playlistmaker.data.settings.SettingsRepo
 import com.example.playlistmaker.domain.settings.model.ThemeList
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
 
-class SettingsRepoImpl(private val sharedPreferences: SharedPreferences?) : SettingsRepo {
 
-    private val PREFS_NAME = "app_settings"
-    private val gson = Gson()
+class SettingsRepoImpl(
+    private val sharedPreferences: SharedPreferences
+) : SettingsRepo {
+
+    companion object {
+        private const val KEY_APPLICATION_THEME = "key_value"
+    }
+
+    init {
+        updateThemeSettings(getThemeSettings())
+    }
 
     override fun getThemeSettings(): ThemeList {
-        val appSettingsString = sharedPreferences?.getString(PREFS_NAME, "")
-        return if (appSettingsString.isNullOrBlank()) {
-            ThemeList.LIGHT
-        } else {
-            try {
-                gson.fromJson(appSettingsString, object : TypeToken<ThemeList>() {}.type)
-            } catch (e: Exception) {
-                Log.e("SETTINGS", "Error reading shared preferences, read error", e);
-                val themeLight = ThemeList.LIGHT
-                updateThemeSettings(themeLight)
-                themeLight
-            }
+        val themeName = sharedPreferences.getString(
+            KEY_APPLICATION_THEME, ThemeList.LIGHT.name
+        ) ?: ThemeList.LIGHT.name
+        return ThemeList.valueOf(themeName)
+    }
 
+    override fun updateThemeSettings(themeList: ThemeList) {
+        val nightMode = when (themeList) {
+            ThemeList.DARK -> AppCompatDelegate.MODE_NIGHT_YES
+            ThemeList.LIGHT -> AppCompatDelegate.MODE_NIGHT_NO
+        }
+        AppCompatDelegate.setDefaultNightMode(nightMode)
+        saveThemeSettings(themeList)
+    }
+
+    private fun saveThemeSettings(applicationTheme: ThemeList) {
+        with(sharedPreferences.edit()) {
+            putString(KEY_APPLICATION_THEME, applicationTheme.name)
+            apply()
         }
     }
-
-    override fun updateThemeSettings(settings: ThemeList) {
-        val appSettingsString = gson.toJson(settings)
-        sharedPreferences?.edit()?.putString(PREFS_NAME, appSettingsString)?.apply()
-    }
-
-    fun saveThemeSettings(settings: ThemeList) {
-        val appSettingsString = gson.toJson(settings)
-        sharedPreferences?.edit()?.putString(PREFS_NAME, appSettingsString)?.apply()
-    }
-
 }
