+++ app/src/main/java/com/example/playlistmaker/domain/medialibrary/impl/MediaLibraryInteractorImpl.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.medialibrary.impl
 
-import com.example.playlistmaker.domain.search.HistoryStorageRepo
+import com.example.playlistmaker.data.search.HistoryStorageRepo
 import com.example.playlistmaker.domain.models.Track
 import com.example.playlistmaker.domain.medialibrary.MediaLibraryInteractor
 class MediaLibraryInteractorImpl(private val historyStorageRepo: HistoryStorageRepo) : MediaLibraryInteractor {
