+++ app/src/main/java/com/example/playlistmaker/domain/settings/impl/SettingsInteractorImpl.kt
@@ -1,13 +1,11 @@
 package com.example.playlistmaker.domain.settings.impl
 
-import com.example.playlistmaker.domain.settings.SettingsRepo
+import com.example.playlistmaker.data.settings.SettingsRepo
 import com.example.playlistmaker.domain.settings.SettingsInteractor
 import com.example.playlistmaker.domain.settings.model.ThemeList
 
 
-class SettingsInteractorImpl(
-    private val settingsRepo: SettingsRepo
-) : SettingsInteractor {
+class SettingsInteractorImpl(private val settingsRepo: SettingsRepo) : SettingsInteractor {
 
     override fun getThemeSettings(): ThemeList = settingsRepo.getThemeSettings()
